+++ CHANGELOG.md
@@ -4,7 +4,8 @@ ChangeLog
 2.2.0-alpha3 (2015-02-25)
 -------------------------
 
-* Also contains all the changes introduced between 2.1.2 and 2.1.3.
+* Contains all the changes introduced between 2.1.2 and 2.1.3.
+* Upgraded to sabre/http 4.0.0alpha1 and now also using sabre/uri.
 * The zip release ships with [sabre/vobject 3.4.2][vobj],
   [sabre/http 4.0.0-alpha1][http], [sabre/event 2.0.1][evnt] and
   [sabre/uri 1.0.0][uri].
@@ -15,6 +16,8 @@ ChangeLog
 
 * Renamed `Sabre\DAV\Auth\Backend\BackendInterface::requireAuth` to
   `challenge`, which is a more correct and better sounding name.
+* The zip release ships with [sabre/vobject 3.3.5][vobj],
+  [sabre/http 3.0.4][http], [sabre/event 2.0.1][evnt].
 
 
 2.2.0-alpha1 (2014-12-10)